Added user functional pages (choosing beautyservice, master and time). Added services and JDBC implementations for master and appointment

+++ src/main/java/net/ukr/lina_chen/model/dao/GenericDao.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface GenericDao<T> extends AutoCloseable {
     void create(T entity) throws SQLException;
-    T findById(int id);
+    T findById(Long id);
     List<T> findAll();
     void update(T entity);
-    void delete(int id);
+    void delete(Long id);
     void close();
 }
